tests/wpt/web-platform-tests/bluetooth/requestDevice/cross-origin-iframe.sub.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<body>
<script>
'use strict';
const test_desc = 'Request device from a unique origin. ' +
    'Should reject with SecurityError.';
const cross_origin_src = 'https://{{domains[www]}}:{{ports[https][0]}}' +
    '/bluetooth/resources/health-thermometer-iframe.html'
let iframe = document.createElement('iframe');

bluetooth_test(async (t) => {
    await setUpHealthThermometerDevice();

    // 1. Load the iframe.
    const iframeWatcher = new EventWatcher(t, iframe, ['load']);
    iframe.src = cross_origin_src;
    document.body.appendChild(iframe);
    await iframeWatcher.wait_for('load');

    // 2. Request the device from the iframe.
    const windowWatcher = new EventWatcher(t, window, ['message']);
    iframe.contentWindow.postMessage({type: 'RequestDevice'}, '*');
    const messageEvent = await windowWatcher.wait_for('message');
    assert_equals(
        messageEvent.data,
        'SecurityError: requestDevice() called from cross-origin iframe.');
}, test_desc);
</script>
</body>
